backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
vosk==0.3.45
torch==2.1.0
transformers==4.35.0
python-multipart==0.0.6
python-dotenv==1.0.0
pyaudio==0.2.13
aiosqlite==0.19.0